+++ frontend/src/App.tsx
@@ -1,12 +1,13 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import LandingPage from './pages/LandingPage'
+import ChatPage from './pages/ChatPage'
 
 function App() {
 	return (
 		<BrowserRouter>
 			<Routes>
 				<Route path='/' element={<LandingPage />} />
-				{/* 更多路由待添加 */}
+				<Route path='/chat' element={<ChatPage />} />
 			</Routes>
 		</BrowserRouter>
 	)